updated the css folder
changed the css styling to the BEM convention as well as created new css files to reflect it.
fixed the heart icon in the image preview modal to display well and also clickable.
ajusted the profile image to sync with the line of the profile info.
Also added some extra accessibility were needed.
Improved the width of the image in the preview stage.